Create new file for project board tests
