+++ spec/javascripts/actions/history_spec.js
@@ -58,26 +58,4 @@ describe('History Actions', () => {
     });
     expect(fakeDispatch).toHaveBeenCalledWith(actions.errorLoadHistory());
   });
-
-  it('Should dispatch toggleHistory when storyId is true', async () => {
-    const storyId = 127;
-
-    const fakeGetHash = sinon.stub().returns(storyId);
-
-    const fakeDispatch = sinon.stub().resolves({});
-
-    await expandStoryIfNeeded(fakeDispatch, fakeGetHash);
-    expect(fakeDispatch).toHaveBeenCalledWith(toggleStory(storyId));
-  })
-
-  it('Should not dispatch toggleHistory when storyId is false', async () => {
-    const storyId = null;
-
-    const fakeGetHash = sinon.stub().returns(storyId);
-
-    const fakeDispatch = sinon.stub().resolves({});
-
-    await expandStoryIfNeeded(fakeDispatch, fakeGetHash);
-    expect(fakeDispatch).toHaveBeenCalledWith(toggleStory(storyId));
-  })
 });
